Created addPhotoViewModel
